Separate joi-extensions from common validators
